chore: moved a bunch of code around to match new plan for builder
