+++ apps/frontend/src/components/atoms/viewTimetable/GeneratingSkeleton.tsx
