+++ apps/frontend/src/components/molecules/viewTimetable/PreferenceRow.tsx
